rename web/beans to web/requests
